# Conflicts:
#	env_list_utils.c
#	lib/minishell.h

+++ parsing/build_lst_utils.c
@@ -13,7 +13,7 @@
 //}
 
 // Helper function to process quoted strings
-void	process_quotes(char *input, int *i, char *buffer, int *buf_index)
+void	process_quotes(char *input, size_t *i, char *buffer, int *buf_index)
 {
 	char	quote_type = input[*i];
 	char*	quote_content;
@@ -37,7 +37,7 @@ void	process_quotes(char *input, int *i, char *buffer, int *buf_index)
 }
 
 // Helper function to process escape sequences
-void	process_escape_sequence(char* input, int* i, char* buffer, int* buf_index)
+void	process_escape_sequence(char *input, size_t *i, char *buffer, int *buf_index)
 {
 	fprintf(stderr, "double slash if \n");
 	buffer[(*buf_index)++] = handle_escape_sequence(input, i, false);
@@ -52,7 +52,7 @@ void	process_escape_sequence(char* input, int* i, char* buffer, int* buf_index)
 	}
 }
 // Helper function to process environment variables
-void	process_env_variable(char* input, int* i, char* buffer, int* buf_index)
+void	process_env_variable(char* input, size_t *i, char* buffer, int* buf_index)
 {
 	fprintf(stderr, "$ if \n");
 	char* env_var;
@@ -66,7 +66,7 @@ void	process_env_variable(char* input, int* i, char* buffer, int* buf_index)
 }
 // Helper function to process regular text
 
-void	process_regular_text(char *input, int *i, char *buffer, int *buf_index)
+void	process_regular_text(char *input, size_t *i, char *buffer, int *buf_index)
 {
 	while (*i < ft_strlen(input) && !ft_isspace(input[*i])
 		&& input[*i] != '|' && input[*i] != '"'
